Fix redis exposed ports
Because of the way Docker deals with firewalls, using "ports" means the
service will be exposed to the outside world, which is not a good thing
for a service like a passwordless redis. Using "expose" directive allows
the other container to see it but it stays confined to the internal
network. Also added a gitignore for docker-compose.yml file and
specified a version for the redis image.